cli.console: refactor and move into sub-package
