+++ src/streamlink_cli/console/console.py
@@ -6,23 +6,7 @@ from getpass import getpass
 from json import dumps
 from typing import Any, TextIO
 
-from streamlink.user_input import UserInputRequester
 from streamlink_cli.utils import JSONEncoder
-
-
-class ConsoleUserInputRequester(UserInputRequester):
-    """
-    Request input from the user on the console using the standard ask/askpass methods
-    """
-
-    def __init__(self, console: ConsoleOutput):
-        self.console = console
-
-    def ask(self, prompt: str) -> str:
-        return self.console.ask(f"{prompt.strip()}: ")
-
-    def ask_password(self, prompt: str) -> str:
-        return self.console.ask_password(f"{prompt.strip()}: ")
 
 
 class ConsoleOutput:
@@ -129,6 +113,3 @@ class ConsoleOutput:
         if self._file_output is not None:
             msg = dumps(out, cls=JSONEncoder, ensure_ascii=False, indent=2)
             self._write(self._file_output, f"{msg}\n")
-
-
-__all__ = ["ConsoleOutput", "ConsoleUserInputRequester"]
